Added Appointment api's

+++ src/main/java/com/hmsapplication/controller/PatientController.java
@@ -24,7 +24,7 @@ public class PatientController {
     }
 
     @GetMapping("/count")
-    public int findPatientCount(){
+    public long findPatientCount(){
         return patientService.getPatientCount();
     }
